+++ CHANGELOG.md
@@ -10,13 +10,18 @@ Change Log
 - “Input in unicode hex” now supports multi-cursor editing.
 
 
+### Fixes
+
+- Improve stability.
+
+
 
 3.7.2 (347)
 --------------------------
 
 ### Improvements
 
-- Make the i-beam cursor more legible in vertical text orientation and dark background theme.
+- Make the i-beam cursor legible in vertical text orientation and dark background theme.
 - Use monospace numbers for line numbers in Find All result table.
 - Optimize performance highlighting found matches.
 - [dev] Update build environment to Xcode 10.2 (Swift 5).