+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Change Log
 
 ### Fixes
 
+- Fix an issue where user settings could not be overwritten to export when the same filename already exists.
 - Improve stability.